Fix and greatly simplify the example app
Perhaps there's something I'm misunderstanding here, but from what I
tell, the example app in this repo has been abandoned, and doesn't
actually work anymore.  There are synax errors that cause it to fail at
runtime, due to the use of `import` in index.html. Furthermore, there is
a preload.js script that contains the same code that is in the
index.html file, but that preload.js file isn't being specified when
creating the `BrowserWindow` in main.js.

This change fixes these issues, and in doing so, it greatly simplifies
the example app.